+++ changelog.md
@@ -1,4 +1,19 @@
 
+## Version 1.10 (2014-08-22)
+
+1. **Added Financial Register Export**
+
+    Added the ability to export the contents of the Actual Income and Expense fields as a financial register containing a series of transactions.
+
+2. **Fixed Delete Logic**
+
+    Fixed a bug that caused problems when deleting an event.
+
+3. **Added Publish Function**
+
+    Added Publish function, which can run a PSTextMerge script to publish the club's events to a planning calendar.
+
+
 ## Version 1.00 (2014-05-26)
 
 1. **Initial Release**